+++ barcodes/models.py
@@ -5,10 +5,11 @@ import uuid as uuid_object
 
 from django.apps import apps
 from django.core.exceptions import ValidationError
-from django.core.urlresolvers import reverse
 from django.db import models
 from django.db.models import Q
 from django.contrib.postgres.fields import ArrayField
+from django.urls import reverse
+
 from projectroles.models import Project
 
 
@@ -21,20 +22,24 @@ def revcomp(s):
 class BarcodeSetManager(models.Manager):
     """Manager for custom table-level BarcodeSet queries"""
 
-    def find(self, search_term, _keywords=None):
-        """Return objects matching the query.
+    def find(self, search_terms, keywords=None):
+        """
+        Return BarcodeSet objects matching the query.
 
-        :param search_term: Search term (string)
+        :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: Python list of BaseFilesfolderClass objects
+        :return: QuerySet of BarcodeSet objects
         """
         objects = super().get_queryset().order_by("name")
-        objects = objects.filter(
-            Q(name__icontains=search_term)
-            | Q(short_name__icontains=search_term)
-            | Q(description__icontains=search_term)
-        )
-        return objects
+        query = Q()
+        for t in search_terms:
+            query = (
+                query
+                | Q(name__icontains=t)
+                | Q(short_name__icontains=t)
+                | Q(description__icontains=t)
+            )
+        return objects.filter(query)
 
 
 #: Generic barcode set.
@@ -65,7 +70,9 @@ class BarcodeSet(models.Model):
     )
 
     #: The project containing this barcode set.
-    project = models.ForeignKey(Project, help_text="Project in which this objects belongs")
+    project = models.ForeignKey(
+        Project, help_text="Project in which this objects belongs", on_delete=models.CASCADE
+    )
 
     #: Full name of the index set
     name = models.CharField(max_length=100, help_text="Full name of the barcode adapter set")
@@ -116,17 +123,20 @@ class BarcodeSet(models.Model):
 class BarcodeSetEntryManager(models.Manager):
     """Manager for custom table-level ``BarcodeSetEntry`` queries"""
 
-    def find(self, search_term, _keywords=None):
-        """Return objects matching the query.
+    def find(self, search_terms, keywords=None):
+        """
+        Return BarcodeSetEntry objects matching the query.
 
-        :param search_term: Search term (string)
+        :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: Python list of BaseFilesfolderClass objects
+        :return: QuerySet of BarcodeSetEntry objects
         """
         objects = super().get_queryset().order_by("name")
-        query = Q(name__icontains=search_term) | Q(sequence__icontains=search_term)
-        if re.match(r"^[acgtnACGTN]+$", search_term):
-            query = query | Q(sequence__icontains=revcomp(search_term))
+        query = Q()
+        for t in search_terms:
+            query = query | Q(name__icontains=t) | Q(sequence__icontains=t)
+            if re.match(r"^[acgtnACGTN]+$", t):
+                query = query | Q(sequence__icontains=revcomp(t))
         objects = objects.filter(query)
         return objects
 
@@ -156,7 +166,9 @@ class BarcodeSetEntry(models.Model):
 
     #: Alternative names for the adapter (e.g., ['01', '1']), uniqueness also applies here in the context of the
     #: ``BarcodeSet``.
-    aliases = ArrayField(models.CharField(max_length=100, db_index=True, unique=False), default=[])
+    aliases = ArrayField(
+        models.CharField(max_length=100, db_index=True, unique=False), default=list
+    )
 
     #: DNA sequence of the barcode.  In the case of dual indexing, use the sequence as for workflow A.
     sequence = models.CharField(max_length=200)