+++ barcodes/plugins.py
@@ -13,7 +13,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     title = "Barcodes"
     urls = urlpatterns
 
-    icon = "barcode"
+    icon = "mdi:barcode"
 
     entry_point_url_id = "barcodes:barcodeset-list"
 
@@ -40,11 +40,13 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     #: Position in plugin ordering
     plugin_ordering = 11
 
-    def search(self, search_term, user, search_type=None, keywords=None):
+    def search(self, search_terms, user, search_type=None, keywords=None):
         """
-        Return app items based on a search term, user, optional type and optional keywords
+        Return app items based on one or more search terms, user, optional type
+        and optional keywords.
 
-        :param search_term: String
+        :param search_terms: Search terms to be joined with the OR operator
+                             (list of strings)
         :param user: User object for user initiating the search
         :param search_type: String
         :param keywords: List (optional)
@@ -58,18 +60,18 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
             projects = Project.objects.all()
 
         if not search_type:
-            barcode_sets = BarcodeSet.objects.find(search_term, keywords).filter(
+            barcode_sets = BarcodeSet.objects.find(search_terms, keywords).filter(
                 project__in=projects
             )
-            barcode_set_entries = BarcodeSetEntry.objects.find(search_term, keywords).filter(
+            barcode_set_entries = BarcodeSetEntry.objects.find(search_terms, keywords).filter(
                 barcode_set__project__in=projects
             )
             items = list(barcode_sets) + list(barcode_set_entries)
             items.sort(key=lambda x: x.name.lower())
         elif search_type == "barcodeset":
-            items = BarcodeSetEntry.objects.find(search_term, keywords)
+            items = BarcodeSetEntry.objects.find(search_terms, keywords)
         elif search_type == "barcodesetentry":
-            items = BarcodeSetEntry.objects.find(search_term, keywords)
+            items = BarcodeSetEntry.objects.find(search_terms, keywords)
 
         return {
             "all": {