+++ digestiflow/__init__.py
@@ -5,3 +5,7 @@ del get_versions
 __version_info__ = tuple(
     [int(num) if num.isdigit() else num for num in __version__.replace("-", ".", 1).split(".")]
 )
+
+from . import _version  # noqa
+
+__version__ = _version.get_versions()["version"]