+++ fileboxes/plugins.py
@@ -13,7 +13,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     title = "File Boxes"
     urls = urlpatterns
 
-    icon = "truck"
+    icon = "mdi:truck"
 
     entry_point_url_id = "fileboxes:filebox-list"
 
@@ -40,11 +40,13 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     #: Position in plugin ordering
     plugin_ordering = 100
 
-    def search(self, search_term, user, search_type=None, keywords=None):
+    def search(self, search_terms, user, search_type=None, keywords=None):
         """
-        Return app items based on a search term, user, optional type and optional keywords
+        Return app items based on one or more search terms, user, optional type
+        and optional keywords.
 
-        :param search_term: String
+        :param search_terms: Search terms to be joined with the OR operator
+                             (list of strings)
         :param user: User object for user initiating the search
         :param search_type: String
         :param keywords: List (optional)
@@ -58,13 +60,19 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
             projects = Project.objects.all()
 
         if not search_type:
-            file_boxes = FileBox.objects.find(search_term, keywords).filter(project__in=projects)
+            file_boxes = FileBox.objects.find(search_terms, keywords).filter(project__in=projects)
             items = list(file_boxes)
             items.sort(key=lambda x: x.title.lower())
         elif search_type == "filebox":
-            items = FileBox.objects.find(search_term, keywords).filter(project__in=projects)
+            items = FileBox.objects.find(search_terms, keywords).filter(project__in=projects)
 
-        return {"all": {"title": "File Boxes", "search_types": ["filebox"], "items": items,}}
+        return {
+            "all": {
+                "title": "File Boxes",
+                "search_types": ["filebox"],
+                "items": items,
+            }
+        }
 
     def get_object_link(self, model_str, uuid):
         """