+++ flowcells/models.py
@@ -3,12 +3,13 @@ import uuid as uuid_object
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
-from django.contrib.postgres.fields import ArrayField, JSONField
+from django.contrib.postgres.fields import ArrayField
 from django.core.exceptions import ValidationError
-from django.core.urlresolvers import reverse
 from django.core.validators import MinValueValidator
 from django.db import models, transaction
-from django.db.models import Q
+from django.db.models import JSONField, Q
+from django.urls import reverse
+
 from mail_factory import factory
 import pagerange
 from projectroles.models import Project, PROJECT_TAG_STARRED
@@ -153,21 +154,25 @@ FLOWCELL_ERROR_CACHE_VERSION = 3
 class FlowCellManager(models.Manager):
     """Manager for custom table-level SequencingMachine queries"""
 
-    def find(self, search_term, keywords=None):
-        """Return objects matching the query.
+    def find(self, search_terms, keywords=None):
+        """
+        Return FlowCell objects matching the query.
 
-        :param search_term: Search term (string)
+        :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: Python list of BaseFilesfolderClass objects
+        :return: QuerySet of FlowCell objects
         """
         objects = super().get_queryset()
-        objects = objects.filter(
-            Q(vendor_id__icontains=search_term)
-            | Q(label__icontains=search_term)
-            | Q(manual_label__icontains=search_term)
-            | Q(description__icontains=search_term)
-        )
-        return objects
+        query = Q()
+        for t in search_terms:
+            query = (
+                query
+                | Q(vendor_id__icontains=t)
+                | Q(label__icontains=t)
+                | Q(manual_label__icontains=t)
+                | Q(description__icontains=t)
+            )
+        return objects.filter(query)
 
 
 def validate_bases_mask(value):
@@ -194,7 +199,9 @@ class FlowCell(models.Model):
     )
 
     #: The project containing this barcode set.
-    project = models.ForeignKey(Project, help_text="Project in which this flow cell belongs")
+    project = models.ForeignKey(
+        Project, help_text="Project in which this flow cell belongs", on_delete=models.CASCADE
+    )
 
     #: Run date of the flow cell
     run_date = models.DateField()
@@ -877,7 +884,11 @@ class FlowCellTag(models.Model):
 
     #: FlowCell to which the tag is assigned
     flowcell = models.ForeignKey(
-        FlowCell, null=False, related_name="tags", help_text="FlowCell to which the tag is assigned"
+        FlowCell,
+        null=False,
+        related_name="tags",
+        help_text="FlowCell to which the tag is assigned",
+        on_delete=models.CASCADE,
     )
 
     #: User for whom the tag is assigned
@@ -886,6 +897,7 @@ class FlowCellTag(models.Model):
         null=False,
         related_name="flowcell_tags",
         help_text="User for whom the tag is assigned",
+        on_delete=models.CASCADE,
     )
 
     #: Name of tag to be assigned
@@ -988,27 +1000,30 @@ REFERENCE_CHOICES = (
 class LibraryManager(models.Manager):
     """Manager for custom table-level Library queries"""
 
-    def find(self, search_term, _keywords=None):
-        """Return objects matching the query.
+    def find(self, search_terms, keywords=None):
+        """
+        Return Library objects matching the query.
 
-        :param search_term: Search term (string)
+        :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: Python list of BaseFilesfolderClass objects
+        :return: QuerySet of Library objects
         """
         objects = super().get_queryset()
-        objects = objects.filter(
-            Q(name__icontains=search_term)
-            | Q(barcode__sequence__icontains=search_term)
-            | Q(barcode_seq__icontains=search_term)
-            | Q(barcode2__sequence__icontains=search_term)
-            | Q(barcode_seq2__icontains=search_term)
-        )
-        return objects
+        query = Q()
+        for t in search_terms:
+            query = (
+                query
+                | Q(name__icontains=t)
+                | Q(barcode__sequence__icontains=t)
+                | Q(barcode_seq__icontains=t)
+                | Q(barcode2__sequence__icontains=t)
+                | Q(barcode_seq2__icontains=t)
+            )
+        return objects.filter(query)
 
 
 class Library(models.Model):
-    """The data stored for each library that is to be sequenced
-    """
+    """The data stored for each library that is to be sequenced"""
 
     #: DateTime of creation
     date_created = models.DateTimeField(auto_now_add=True, help_text="DateTime of creation")
@@ -1195,16 +1210,19 @@ FORMAT_CHOICES = ((FORMAT_PLAIN, "Plain Text"), (FORMAT_MARKDOWN, "Markdown"))
 class MessageManager(models.Manager):
     """Manager for custom table-level Message queries"""
 
-    def find(self, search_term, _keywords=None):
-        """Return objects matching the query.
+    def find(self, search_terms, keywords=None):
+        """
+        Return Message objects matching the query.
 
-        :param search_term: Search term (string)
+        :param search_terms: Search terms (list of strings)
         :param keywords: Optional search keywords as key/value pairs (dict)
-        :return: Python list of BaseFilesfolderClass objects
+        :return: QuerySet of Message objects
         """
         objects = super().get_queryset()
-        objects = objects.filter(Q(subject__icontains=search_term) | Q(body__icontains=search_term))
-        return objects
+        query = Q()
+        for t in search_terms:
+            query = query | Q(subject__icontains=t) | Q(body__icontains=t)
+        return objects.filter(query)
 
 
 class Message(models.Model):
@@ -1357,7 +1375,7 @@ def message_created(message):
 
 
 class KnownIndexContamination(models.Model):
-    """Known contamination """
+    """Known contamination"""
 
     #: DateTime of creation
     date_created = models.DateTimeField(auto_now_add=True, help_text="DateTime of creation")